Creamos la carpeta vista
