+++ src/co/edu/udea/MVCDAO/negocio/Negocio.java
@@ -1,0 +1,5 @@
+package co.edu.udea.MVCDAO.negocio;
+
+public class Negocio {
+
+}
