+++ src/co/edu/udea/MVCDAO/vista/VentanaIngresar.java
@@ -1,0 +1,168 @@
+/*
+ * Created by JFormDesigner on Mon Dec 07 12:09:18 COT 2020
+ */
+
+package co.edu.udea.MVCDAO.vista;
+
+import java.awt.*;
+import java.awt.event.*;
+import javax.swing.*;
+
+import co.edu.udea.MVCDAO.dao.impl.EstudianteDAOFILE;
+import co.edu.udea.MVCDAO.modelo.EstudianteDTO;
+import com.intellij.uiDesigner.core.*;
+
+/**
+ * @author Diego Munoz
+ */
+public class VentanaIngresar extends JFrame {
+    public VentanaIngresar() {
+        initComponents();
+    }
+
+    EstudianteDAOFILE accesoArchivo = new EstudianteDAOFILE();
+
+    private void botonIngresar_ingresarMouseClicked(MouseEvent e) {
+        String documento = textField4.getText();
+        String nombre = textField1.getText();
+        String genero = textField3.getText();
+        if (nombre.equalsIgnoreCase("") || documento.equalsIgnoreCase("") || genero.equalsIgnoreCase("")) {
+            JOptionPane.showMessageDialog(this, "Datos incompletos");
+        } else {
+            EstudianteDTO estudiante = new EstudianteDTO(nombre, textField2.getText(), genero.charAt(0), textField4.getText());
+            accesoArchivo.almacenarEstudiante(estudiante);
+            JOptionPane.showMessageDialog(this, "Guardado satisfactoriamente");
+            //this.dispose();
+            textField1.setText("");
+            textField2.setText("");
+            textField3.setText("");
+            textField4.setText("");
+
+        }
+    }
+
+    private void botonIngresarMouseClicked(MouseEvent e) {
+        // TODO add your code here
+    }
+
+    private void button1MouseClicked(MouseEvent e) {
+        this.dispose();
+    }
+
+    private void initComponents() {
+        // JFormDesigner - Component initialization - DO NOT MODIFY  //GEN-BEGIN:initComponents
+        // Generated using JFormDesigner Evaluation license - unknown
+        label1 = new JLabel();
+        textField1 = new JTextField();
+        label2 = new JLabel();
+        textField2 = new JTextField();
+        label3 = new JLabel();
+        textField3 = new JTextField();
+        label4 = new JLabel();
+        textField4 = new JTextField();
+        botonIngresar_ingresar = new JButton();
+        button1 = new JButton();
+
+        //======== this ========
+        Container contentPane = getContentPane();
+
+        //---- label1 ----
+        label1.setText("Nombre");
+
+        //---- label2 ----
+        label2.setText("Apellido");
+
+        //---- label3 ----
+        label3.setText("Genero");
+
+        //---- label4 ----
+        label4.setText("Documento");
+
+        //---- botonIngresar_ingresar ----
+        botonIngresar_ingresar.setText("Ingresar");
+        botonIngresar_ingresar.addMouseListener(new MouseAdapter() {
+            @Override
+            public void mouseClicked(MouseEvent e) {
+                botonIngresarMouseClicked(e);
+                botonIngresar_ingresarMouseClicked(e);
+            }
+        });
+
+        //---- button1 ----
+        button1.setText("Salir");
+        button1.addMouseListener(new MouseAdapter() {
+            @Override
+            public void mouseClicked(MouseEvent e) {
+                button1MouseClicked(e);
+            }
+        });
+
+        GroupLayout contentPaneLayout = new GroupLayout(contentPane);
+        contentPane.setLayout(contentPaneLayout);
+        contentPaneLayout.setHorizontalGroup(
+                contentPaneLayout.createParallelGroup()
+                        .addGroup(contentPaneLayout.createSequentialGroup()
+                                .addGap(33, 33, 33)
+                                .addGroup(contentPaneLayout.createParallelGroup()
+                                        .addComponent(label2)
+                                        .addComponent(label1, GroupLayout.PREFERRED_SIZE, 55, GroupLayout.PREFERRED_SIZE)
+                                        .addComponent(label3)
+                                        .addComponent(label4))
+                                .addGap(59, 59, 59)
+                                .addGroup(contentPaneLayout.createParallelGroup()
+                                        .addComponent(textField2, GroupLayout.PREFERRED_SIZE, 201, GroupLayout.PREFERRED_SIZE)
+                                        .addComponent(textField1, GroupLayout.PREFERRED_SIZE, 201, GroupLayout.PREFERRED_SIZE)
+                                        .addComponent(textField3, GroupLayout.PREFERRED_SIZE, 201, GroupLayout.PREFERRED_SIZE)
+                                        .addComponent(textField4, GroupLayout.PREFERRED_SIZE, 201, GroupLayout.PREFERRED_SIZE))
+                                .addContainerGap(22, Short.MAX_VALUE))
+                        .addGroup(GroupLayout.Alignment.TRAILING, contentPaneLayout.createSequentialGroup()
+                                .addContainerGap(97, Short.MAX_VALUE)
+                                .addComponent(botonIngresar_ingresar)
+                                .addGap(35, 35, 35)
+                                .addComponent(button1)
+                                .addGap(90, 90, 90))
+        );
+        contentPaneLayout.setVerticalGroup(
+                contentPaneLayout.createParallelGroup()
+                        .addGroup(contentPaneLayout.createSequentialGroup()
+                                .addGap(36, 36, 36)
+                                .addGroup(contentPaneLayout.createParallelGroup(GroupLayout.Alignment.BASELINE)
+                                        .addComponent(textField1, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
+                                        .addComponent(label1))
+                                .addGap(31, 31, 31)
+                                .addGroup(contentPaneLayout.createParallelGroup(GroupLayout.Alignment.BASELINE)
+                                        .addComponent(textField2, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
+                                        .addComponent(label2))
+                                .addGap(29, 29, 29)
+                                .addGroup(contentPaneLayout.createParallelGroup(GroupLayout.Alignment.BASELINE)
+                                        .addComponent(textField3, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
+                                        .addComponent(label3))
+                                .addGap(28, 28, 28)
+                                .addGroup(contentPaneLayout.createParallelGroup(GroupLayout.Alignment.BASELINE)
+                                        .addComponent(textField4, GroupLayout.PREFERRED_SIZE, GroupLayout.DEFAULT_SIZE, GroupLayout.PREFERRED_SIZE)
+                                        .addComponent(label4))
+                                .addPreferredGap(LayoutStyle.ComponentPlacement.RELATED, 30, Short.MAX_VALUE)
+                                .addGroup(contentPaneLayout.createParallelGroup(GroupLayout.Alignment.BASELINE)
+                                        .addComponent(botonIngresar_ingresar, GroupLayout.PREFERRED_SIZE, 52, GroupLayout.PREFERRED_SIZE)
+                                        .addComponent(button1, GroupLayout.PREFERRED_SIZE, 52, GroupLayout.PREFERRED_SIZE))
+                                .addGap(17, 17, 17))
+        );
+        pack();
+        setLocationRelativeTo(getOwner());
+        // JFormDesigner - End of component initialization  //GEN-END:initComponents
+    }
+
+    // JFormDesigner - Variables declaration - DO NOT MODIFY  //GEN-BEGIN:variables
+    // Generated using JFormDesigner Evaluation license - unknown
+    private JLabel label1;
+    private JTextField textField1;
+    private JLabel label2;
+    private JTextField textField2;
+    private JLabel label3;
+    private JTextField textField3;
+    private JLabel label4;
+    private JTextField textField4;
+    private JButton botonIngresar_ingresar;
+    private JButton button1;
+    // JFormDesigner - End of variables declaration  //GEN-END:variables
+}
